Code Cleanup

+++ app/src/main/java/com/ugm/kaskita/utils/Converter.kt
@@ -1,36 +1,35 @@
 package com.ugm.kaskita.utils
 
-import java.text.DecimalFormat
 import java.text.NumberFormat
 import java.util.*
 
 object Converter {
-    
+
     fun String.formatRupiah(): String {
         val localeID = Locale("in", "ID")
         val format =
             NumberFormat.getCurrencyInstance(localeID)
         return format.format(this.toLong())
-    }    
-    
+    }
+
     fun Int.formatRupiah(): String {
         val localeID = Locale("in", "ID")
         val format =
             NumberFormat.getCurrencyInstance(localeID)
         return format.format(this.toLong())
-    }    
-    
+    }
+
     fun Float.formatRupiah(): String {
         val localeID = Locale("in", "ID")
         val format =
             NumberFormat.getCurrencyInstance(localeID)
         return format.format(this.toLong())
-    }    
-    
+    }
+
     fun Long.formatRupiah(): String {
         val localeID = Locale("in", "ID")
         val format =
             NumberFormat.getCurrencyInstance(localeID)
-        return format.format(this.toLong())
+        return format.format(this)
     }
 }
